requirements-frozen.txt
-f https://pypi.research.pdx.edu/dist/
psu.oit.wdt.ecoroofs==1.0.0
psu.oit.arc.tasks
appdirs==1.4.3
boto3==1.4.4
botocore==1.5.44
certifi==2017.4.17
click==6.7
configobj==5.0.6
contextlib2==0.5.5
Django==1.10.7
django-arcutils==2.17
django-local-settings==1.0b6
django-pgcli==0.0.2
djangorestframework==3.6.2
docutils==0.13.1
flake8==3.3.0
humanize==0.5.1
jmespath==0.9.2
Markdown==2.6.8
mccabe==0.6.1
packaging==16.8
pgcli==1.5.1
pgspecial==1.7.0
prompt-toolkit==1.0.14
psycopg2==2.7.1
pycodestyle==2.3.1
pyflakes==1.5.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
raven==6.0.0
-e git+git@github.com:wylee/runcommands.git@4113e11df8d53ad1f5a3f8538e32d2f6c383287f#egg=runcommands
s3transfer==0.1.10
setproctitle==1.1.10
six==1.10.0
sqlparse==0.2.3
stashward==0.1.4
Unidecode==0.4.20
wcwidth==0.1.7
